refactor: seperate layout.tsx
